+++ src/main/java/lotto/util/AmountValidator.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.util;
 
 import lotto.exception.AmountErrorMessage;
 import lotto.exception.LottoConstants;
