+++ src/main/java/lotto/util/CountConverter.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.util;
 
 import lotto.exception.LottoConstants;
